requirements.txt
Pillow
numpy
pandas
scikit-learn
torch
torchvision
tqdm
